refactor: 디렉토리 /global/dto를 /global/error로 수정
