+++ server/src/main/java/ppalatjyo/server/global/security/handler/CustomAccessDeniedHandler.java
@@ -9,7 +9,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.web.access.AccessDeniedHandler;
-import ppalatjyo.server.global.dto.error.ErrorResponseDto;
+import ppalatjyo.server.global.error.ErrorResponseDto;
 
 import java.io.IOException;
